refactor(context-menu): extract shared GTK renderer
Move the duplicated GTK popover menu rendering logic into one shared infrastructure package so CEF and WebKit keep using the same tested menu behavior.
